tools: splitting flags and proto scripts into separae files
